fix!: prevent timestamps shifting by an hour whenever a file was changed
BREAKING CHANGE: Timestamps are now stored as seconds from the UNIX epoch so
they can be converted straight to a time_t.

Storing timestamps as a human readable string was causing issues as timezone and
GMT offset were not being decoded properly (C++17 does not have an efficient way
of reading in the timezone information from a string).